+++ lib/SXNGN/headers/Database.h
@@ -6,17 +6,17 @@ namespace SXNGN {
     class Database {
     public:
         Database();
-        static int scale_;
-        static int get_scale();
-        static void set_scale(int new_scale);
+        static Uint8 scale_;
+        static Uint8 get_scale();
+        static void set_scale(Uint8 new_scale);
         static void reduce_scale();
         static void increase_scale();
 
-        static int max_fps_;
-        static int get_max_fps();
-        static void set_max_fps(int new_fps);
+        static Uint8 max_fps_;
+        static Uint8 get_max_fps();
+        static void set_max_fps(Uint8 new_fps);
 
-        static int get_screen_ticks_per_frame();
+        static Uint8 get_screen_ticks_per_frame();
 
         
         static std::shared_ptr<Coordinator> ecs_coordinator_;
